In calendar, add all the events of the current week.
Currently only two events are written in the calendar to be downloaded; usually this would work just fine, but during special events (like this week) where there are multiple turtle times per week, it is not possible to import all of them.

With this fix all the events of the current week will be added to the .ics file. In the future we may think of a UI widget to allow users to specify on how many days they want the ics to be created.